Remove inapplicable TODO (#33)
It is not possible for the stream to respect back pressure since
URLSessionDelegate methods are called by URLSession without a way to
delay them (other than blocking the delegate thread which just moves the
buffer elsewhere).

For data delivery, we are suspending and resuming the task based on how
much data remains to be consumed. But for events we just have to buffer
them.